+++ src/ppl/kernel/llm/cuda/xformer/fmha.cu
@@ -1,5 +1,22 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
-#include "fmha.h"
+#include "ppl/kernel/llm/cuda/xformer/fmha.h"
+#include "utils/kernel_forward.h"
 
 #include "autogen/cutlassF.h"
 
@@ -38,7 +55,7 @@ ppl::common::RetCode fmha(
     const int64_t num_kv_heads,
     const int64_t head_dim,
     const int64_t custom_mask_type,
-    const double attn_scale,
+    const float attn_scale,
     void* output)
 {
     if (datatype != ppl::common::DATATYPE_FLOAT16) {
@@ -105,9 +122,9 @@ ppl::common::RetCode fmha(
         p.seqlen_k_ptr = nullptr;
 
         if (attn_scale != 0) {
-            p.scale = float(attn_scale);
+            p.scale = attn_scale;
         } else {
-            p.scale = float(1.0 / std::sqrt(float(p.head_dim)));
+            p.scale = 1.0f / std::sqrt(float(p.head_dim));
         }
 
         p.q_strideB = query_stride_b;
